Added new attack type, worked on previous attacks
Added new attack type: claw attack - summons a hand when player in range. It will deal damage and push player away. Needs a bit of overwork because it is binded to the shield that is constantly moving up and down, so the shield has to move to certain location so that the hand could reach the player. Refactored laser code, now it will be easy to change properties of laser attack for further boss stages